refactor: 메서드 반환형 int->Long으로 수정

+++ domain/src/main/java/org/depromeet/spot/domain/review/ReviewListResult.java
@@ -5,6 +5,6 @@ import java.util.List;
 public record ReviewListResult(
         List<Review> reviews,
         List<KeywordCount> topKeywords,
-        int totalCount,
+        Long totalCount,
         int offset,
         int limit) {}
